requirements.txt
flask
flask-socketio
eventlet
opencv-python
tensorflow
gTTS
google-generativeai
